# swagger/spec.yml
openapi: 3.0.3
info:
  title: API Калькулятор расчета стоимости доставки
  version: 1.0.0
  description: |
    *Калькулятор расчета стоимости доставки* - сервис расчета доставки
    ### Возможности Сервиса
      - Расчет стоимостей
      - Регистр месяцев
      - Регистр тоннажей
      - Регистр стоимости доставки
    
    Разработчик [Андрей] (https://gitlab-dev.efko.ru/cr-ovr5)
...
paths:
...
  /api/v1/price:
    post:
      tags:
        - price
      summary: Добавить цену
      requestBody:
        $ref: '#/components/requestBodies/PriceRequestBody'
      responses:
        '201':
          description: Успешное добавление
        '401':
          description: Не авторизован

    put:
      tags:
        - price
      summary: Обновить цену
      requestBody:
        $ref: '#/components/requestBodies/PriceRequestBody'
      responses:
        '204':
          description: Успешное удаление
        '401':
          description: Не авторизован
components:
  requestBodies:
...
    PriceRequestBody:
      description: Данные по стоимости
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: "Шрот"
              month:
                type: string
                example: "Март"
              tonnage:
                type: integer
                example: 100
              value:
                type: integer
                example: 176
...